+++ src/main/java/com/umc7/ZIC/security/OAuth/CustomOAuth2User.java
@@ -1,4 +1,4 @@
-package com.umc7.ZIC.security;
+package com.umc7.ZIC.security.OAuth;
 
 import com.umc7.ZIC.user.domain.User;
 import org.springframework.security.core.GrantedAuthority;
